specify full path for aria2c
I forgot to document and fix an issue I was to tired to commit last night. I was having issues with the script running and deleting the magnet file and leaving me no .torrent. I have to specify the full path to aria2c using the crontab entry I have in the setup file.